agents/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "productivity_intelligence"
version = "0.1.0"
description = "Multi-agent productivity intelligence system with Email, Notion & Calendar agents"
dependencies = [
    "dagster>=1.5.0",
    "dagster-webserver",
    "langchain>=0.1.0",
    "langgraph>=0.1.0", 
    "langchain-google-genai>=1.0.0",
    "google-generativeai>=0.3.0",
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "notion-client>=2.0.0",
    "pandas",
    "requests"
]

[project.optional-dependencies]
dev = [
    "dagster-webserver",
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy"
]

[tool.dagster]
module_name = "agents.definitions"